Avoid non-ASCII in __repr__ (the REPLs don't like it).
As a UI trick, the __repr__ for API call just yields a readable string
that we want to display to the user (and not a Python expression).  It
would be nice if we could display Unicode characters correctly, but both
the standard Python REPL and IPython don't like it when __repr__ returns
non-ASCII characters.